+++ lib/src/model/enum.dart
@@ -1,0 +1,5 @@
+abstract class Enum<T> {
+  const Enum(this.value);
+
+  final T value;
+}