+++ CHANGELOG.md
@@ -1,5 +1,23 @@
 # Changelog
 
+## 1.0.0
+### Added
+- Add Dart 2 compatibility.
+- Add `bool random` argument to `getUsers` function to get random users from the data source.
+- Add custom Gender enumerator.
+- Add `getUser` function to retrieve a single user.
+- Add argument validation to `getUsers` function.
+- Add documentation to `getUsers` function.
+- 
+### Changed
+- Add extra info to README.
+- Change `RandomUser`'s `userName` field to `username`.
+- Change RandomUser.gender `String` to a custom `Gender` enumerator value.
+- Fixed up various spelling errors.
+
+### Removed
+- Removed `dart_dev` tooling dependencies for Dart 2 compatibility.
+
 ## 0.0.3
 
 - Adding a default constructor to `RandomUser`